Move interpretable logic to InterpretableTabularPredictor
