+++ tabular/src/autogluon/tabular/predictor/__init__.py
@@ -1,1 +1,2 @@
+from .interpretable_predictor import InterpretableTabularPredictor
 from .predictor import TabularPredictor
